+++ onyxia-api/src/main/java/fr/insee/onyxia/api/services/AppsService.java
@@ -1,11 +1,15 @@
 package fr.insee.onyxia.api.services;
 
-import java.util.List;
-import java.util.Map;
+import fr.insee.onyxia.model.User;
+import fr.insee.onyxia.model.service.Service;
+import mesosphere.marathon.client.model.v2.Group;
 
-import mesosphere.marathon.client.model.v2.VersionedApp;
+import java.io.IOException;
+import java.util.List;
+import java.util.concurrent.TimeoutException;
 
 public interface AppsService {
-   
-   VersionedApp getServiceById(String id);
+
+    Group getGroups(String id) throws IOException;
+    List<Service> getUserServices(User user) throws InterruptedException, TimeoutException, IOException;
 }
